remove: 임시 타입 파일 제거
