+++ src/mocks/handlers/expense.ts
@@ -3,7 +3,7 @@ import {
   Expense,
   ExpenseForm,
   SingleExpenseForm,
-} from '@/pages/createBill/types/expense.type2';
+} from '@/pages/createBill/types/expense.type';
 import getIsMocked from '@/mocks/utils/getIsMocked';
 import { dummyGroupMembers } from './groupMember';
