+++ src/pages/createBill/utils/getTotalExpense.ts
@@ -1,7 +1,7 @@
 import {
   SingleExpenseForm,
   Expense,
-} from '@/pages/createBill/types/expense.type2';
+} from '@/pages/createBill/types/expense.type';
 
 /**
  * 모든 지출의 누적 합을 반환합니다.
